+++ fluxo-kmp-conf/src/main/kotlin/fluxo/vc/TomlResultCatalog.kt
@@ -14,7 +14,9 @@ import org.gradle.api.artifacts.MinimalExternalModuleDependency
 import org.gradle.api.internal.artifacts.DefaultModuleIdentifier
 import org.gradle.api.internal.artifacts.dependencies.DefaultMinimalDependency
 import org.gradle.api.internal.artifacts.dependencies.DefaultMutableVersionConstraint
+import org.gradle.api.internal.artifacts.dependencies.DefaultPluginDependency
 import org.gradle.api.logging.Logger
+import org.gradle.plugin.use.PluginDependency
 import org.tomlj.TomlParseResult
 import org.tomlj.TomlTable
 
@@ -32,21 +34,72 @@ internal class TomlResultCatalog(
 ) {
     internal val versions: Map<String, String>
     internal val libraries: Map<String, MinimalExternalModuleDependency>
+    internal val plugins: Map<String, PluginDependency>
 
     init {
         val logger = project.logger
         versions = parseVersions(toml.getTable(VERSIONS_KEY), logger)
         libraries = parseLibraries(toml.getTable(LIBRARIES_KEY), logger)
+        plugins = parsePlugins(toml.getTable(PLUGINS_KEY), logger)
 
-        // TODO: Parse bundles and plugins
+        // TODO: Parse bundles
 
         // val bundles = toml.getTable(BUNDLES_KEY)
-        // val plugins = toml.getTable(PLUGINS_KEY)
 
         logger.d(
             "Loaded bundled toml version catalog" +
-                " (${versions.size} versions, ${libraries.size} libraries)",
+                " (${versions.size} versions, " +
+                "${libraries.size} libraries, " +
+                "${plugins.size} plugins)",
         )
+    }
+
+    /** @see org.gradle.api.internal.catalog.parser.TomlCatalogFileParser.parsePlugin */
+    private fun parsePlugins(
+        table: TomlTable?,
+        logger: Logger,
+    ): Map<String, PluginDependency> {
+        if (table == null || table.isEmpty) {
+            return emptyMap()
+        }
+
+        val plugins = HashMap<String, PluginDependency>(table.size())
+        for (alias in table.keySet()) {
+            // short-notation = "some.plugin.id:1.4"
+            // long-notation = { id = "some.plugin.id", version = "1.4" }
+            // reference-notation = { id = "some.plugin.id", version.ref = "common" }
+            plugins[alias] = when (val entry = table.get(alias)) {
+                is String -> {
+                    val (id, version) = entry.split(':', limit = 2)
+                    val v = DefaultMutableVersionConstraint.withVersion(version)
+                    DefaultPluginDependency(id, v)
+                }
+
+                is TomlTable -> {
+                    val entryTable: TomlTable = entry
+                    val id = entryTable.getString("id")
+                    if (id == null) {
+                        logger.e("Missing 'id' for plugin alias '$alias': $entry")
+                        continue
+                    }
+
+                    val version = parseVersion(
+                        v = entryTable.get("version"),
+                        logger = logger,
+                        target = "plugin alias '$alias'",
+                    ) ?: continue
+
+                    val v = DefaultMutableVersionConstraint.withVersion(version)
+                    DefaultPluginDependency(id, v)
+                }
+
+                else -> {
+                    logger.e("Invalid plugin entry for alias '$alias': $entry")
+                    continue
+                }
+            }
+        }
+        return plugins
     }
 
     /** @see org.gradle.api.internal.catalog.parser.TomlCatalogFileParser.parseLibrary */
@@ -84,34 +137,14 @@ internal class TomlResultCatalog(
                         DefaultModuleIdentifier.newId(group, name)
                     }
 
-                    val version = when (val v = entryTable.get("version")) {
-                        is String -> {
-                            DefaultMutableVersionConstraint.withVersion(v)
-                        }
+                    val version = parseVersion(
+                        v = entryTable.get("version"),
+                        logger = logger,
+                        target = "library alias '$alias'",
+                    ) ?: continue
 
-                        is TomlTable -> {
-                            val ref = v.getString("ref")
-                            if (ref == null) {
-                                logger.e("Missing version 'ref' for library alias '$alias': $v")
-                                continue
-                            }
-                            val version = versions[ref]
-                            if (version == null) {
-                                logger.e(
-                                    "Missing referenced version for library alias '$alias': $ref",
-                                )
-                                continue
-                            }
-                            DefaultMutableVersionConstraint.withVersion(version)
-                        }
-
-                        else -> {
-                            logger.e("Invalid version entry for library alias '$alias': $v")
-                            continue
-                        }
-                    }
-
-                    DefaultMinimalDependency(module, version)
+                    val v = DefaultMutableVersionConstraint.withVersion(version)
+                    DefaultMinimalDependency(module, v)
                 }
 
                 else -> {
@@ -131,21 +164,56 @@ internal class TomlResultCatalog(
 
         val versions = HashMap<String, String>(table.size())
         for (alias in table.keySet()) {
-            when (val version = table.get(alias)) {
-                is String -> if (version.isNotEmpty()) {
-                    versions[alias] = version
-                }
-
-                else -> logger.e("Invalid version entry for bundled TOML alias '$alias': $version")
-            }
+            versions[alias] = parseVersion(
+                v = table.get(alias),
+                logger = logger,
+                target = "version alias '$alias'",
+                versions = versions,
+            ) ?: continue
         }
         return versions
     }
 
+
+    @Suppress("ReturnCount")
+    private fun parseVersion(
+        v: Any?,
+        logger: Logger,
+        target: String,
+        versions: Map<String, String> = this.versions,
+    ): String? {
+        return when (v) {
+            is String -> v
+
+            is TomlTable -> {
+                val ref = v.getString("ref")
+                if (ref == null) {
+                    logger.e("Missing version 'ref' for $target: $v")
+                    return null
+                }
+                val version = versions[ref]
+                if (version == null) {
+                    logger.e(
+                        "Missing referenced version for $target: $ref",
+                    )
+                    return null
+                }
+                version
+            }
+
+            else -> {
+                logger.e("Invalid version entry for $target: $v")
+                null
+            }
+        }
+    }
+
+
     private companion object {
         private const val VERSIONS_KEY = "versions"
         private const val LIBRARIES_KEY = "libraries"
-//        private const val BUNDLES_KEY = "bundles"
-//        private const val PLUGINS_KEY = "plugins"
+
+        //        private const val BUNDLES_KEY = "bundles"
+        private const val PLUGINS_KEY = "plugins"
     }
 }